backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pymongo==4.7.2
python-dotenv==1.0.1
gunicorn==21.2.0